Extracted PreviewCoordinator from ZoopController
